Fixing test failures / added test failing on renaming module.

+++ plugins/com.python.pydev.refactoring/tests/com/python/pydev/refactoring/refactorer/ClassHierarchySearchTest.java
@@ -10,11 +10,14 @@
 package com.python.pydev.refactoring.refactorer;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.List;
 
+import org.eclipse.core.runtime.CoreException;
 import org.eclipse.jface.text.Document;
 import org.python.pydev.core.ModulesKey;
 import org.python.pydev.core.docutils.PySelection;
+import org.python.pydev.core.log.Log;
 import org.python.pydev.editor.codecompletion.revisited.ProjectModulesManager;
 import org.python.pydev.editor.codecompletion.revisited.modules.CompiledModule;
 import org.python.pydev.editor.codecompletion.revisited.modules.SourceModule;
@@ -43,21 +46,59 @@ public class ClassHierarchySearchTest extends AdditionalInfoTestsBase {
     }
 
     private Refactorer refactorer;
-    private File baseDir;
+    private static File baseDir;
+    private static File baseDir2;
 
+    @Override
     public void setUp() throws Exception {
-        super.setUp();
         CompiledModule.COMPILED_MODULES_ENABLED = true;
-        this.restorePythonPath(false);
         refactorer = new Refactorer();
-        baseDir = FileUtils.getTempFileAt(new File("."), "data_temp_class_hierarchy_search_test");
-        if (baseDir.exists()) {
-            FileUtils.deleteDirectoryTree(baseDir);
+        if (baseDir != null && !baseDir.exists()) {
+            baseDir.mkdirs();
         }
-        baseDir.mkdir();
+        if (baseDir2 != null && !baseDir2.exists()) {
+            baseDir2.mkdirs();
+        }
+        super.setUp();
+
         SourceModule.TESTING = true;
     }
 
+    @Override
+    public String getProjectPythonpath() {
+        if (baseDir == null) {
+            //This will be called only once for the class and we want to use the same path over and over...
+            baseDir = FileUtils.getTempFileAt(new File("."), "data_temp_class_hierarchy_search_test");
+            if (baseDir.exists()) {
+                try {
+                    FileUtils.deleteDirectoryTree(baseDir);
+                } catch (IOException e) {
+                    Log.log(e);
+                }
+            }
+            baseDir.mkdir();
+        }
+        return super.getProjectPythonpath() + "|" + baseDir.getAbsolutePath();
+    }
+
+    @Override
+    public String getProjectPythonpathNature2() {
+        if (baseDir2 == null) {
+            //This will be called only once for the class and we want to use the same path over and over...
+            baseDir2 = FileUtils.getTempFileAt(new File("."), "data_temp_class_hierarchy_search_test");
+            if (baseDir2.exists()) {
+                try {
+                    FileUtils.deleteDirectoryTree(baseDir2);
+                } catch (IOException e) {
+                    Log.log(e);
+                }
+            }
+            baseDir2.mkdir();
+        }
+        return super.getProjectPythonpathNature2() + "|" + baseDir2.getAbsolutePath();
+    }
+
+    @Override
     public void tearDown() throws Exception {
         CompiledModule.COMPILED_MODULES_ENABLED = false;
         ProjectModulesManager projectModulesManager = ((ProjectModulesManager) nature.getAstManager()
@@ -71,9 +112,9 @@ public class ClassHierarchySearchTest extends AdditionalInfoTestsBase {
         projectModulesManager.doRemoveSingleModule(new ModulesKey("fooIn2", null));
         projectModulesManager.doRemoveSingleModule(new ModulesKey("fooIn20", null));
 
-        if (baseDir.exists()) {
-            FileUtils.deleteDirectoryTree(baseDir);
-        }
+        FileUtils.deleteDirectoryTree(baseDir);
+        FileUtils.deleteDirectoryTree(baseDir2);
+
         super.tearDown();
     }
 
@@ -299,7 +340,8 @@ public class ClassHierarchySearchTest extends AdditionalInfoTestsBase {
 
     private RefactoringRequest setUpModule(final int line, final int col, String str, String modName,
             PythonNature natureToAdd) {
-        File f = new File(baseDir, modName +
+
+        File f = new File(natureToAdd == nature2 ? baseDir2 : baseDir, modName +
                 ".py");
 
         Document doc = new Document(str);
@@ -328,9 +370,14 @@ public class ClassHierarchySearchTest extends AdditionalInfoTestsBase {
                 }
             }
         }
-        fail("Unable to find node with name:" + name +
-                " mod:" + modName +
-                "\nAvailable:" + available);
+        try {
+            fail("Unable to find node with name:" + name +
+                    " mod:" + modName +
+                    "\nAvailable:" + available + "\n\nPythonpath: "
+                    + nature.getPythonPathNature().getOnlyProjectPythonPathStr(true));
+        } catch (CoreException e) {
+            throw new RuntimeException(e);
+        }
         return null;
     }
 }
